Making go lint happy.
