+++ wlnr/relayinterface/client_rpc.go
@@ -1,4 +1,4 @@
-package relay_interface
+package relayinterface
 
 import (
 	"log"
@@ -7,16 +7,22 @@ import (
 	"time"
 )
 
+// ClientRPC is an internal struct which implements relayinterface.Client
+// over a RPC connection.
 type ClientRPC struct {
 	callback ClientCallback
 	relay    *rpc.Client
 	listener net.Listener
 }
 
+// ClientRPCMethods is a helper struct so only some methods are exposed to RPC.
 type ClientRPCMethods struct {
 	client *ClientRPC
 }
 
+// NewClientRPC creates a struct that implements relayinterface.Client over RPC.
+// An RPC server running on localhost:7398 is assumed.
+// Methods of the given callback are called with notifications of the server.
 func NewClientRPC(callback ClientCallback) Client {
 	client := &ClientRPC{
 		callback: callback,
@@ -45,8 +51,8 @@ func NewClientRPC(callback ClientCallback) Client {
 	return client
 }
 
+// Open connection to relay server
 func (client *ClientRPC) connect() bool {
-	// Open connection to relay server
 	connection, err := net.DialTimeout("tcp", "localhost:7398", time.Duration(10)*time.Second)
 	if err != nil {
 		log.Fatal("Unable to connect to relay server at localhost: ", err)
@@ -57,6 +63,7 @@ func (client *ClientRPC) connect() bool {
 	return true
 }
 
+// CloseConnection terminates the connection to the relay server.
 func (client *ClientRPC) CloseConnection() {
 	client.listener.Close()
 }
@@ -89,11 +96,13 @@ func (client *ClientRPC) CreateGame(name string, hostPassword string) bool {
 	return success
 }
 
+// GameConnected is called by the relay over rpc when a host connected to a game.
 func (client *ClientRPCMethods) GameConnected(in *GameData, response *bool) (err error) {
 	client.client.callback.GameConnected(in.Name)
 	return nil
 }
 
+// GameClosed is called by the relay over rpc when a game has ended.
 func (client *ClientRPCMethods) GameClosed(in *GameData, response *bool) (err error) {
 	client.client.callback.GameClosed(in.Name)
 	return nil
